services/mqtt_kafka_connector/requirements.txt
paho-mqtt==1.6.1
kafka-python==2.0.2
python-json-logger==2.0.7
python-snappy==0.6.1